// src/app/generated/dcc/model/dataDto.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ListDto } from './listDto';
import { QuantityDto } from './quantityDto';
import { RichContentDto } from './richContentDto';
import { FormulaDto } from './formulaDto';
import { ByteDataDto } from './byteDataDto';


export interface DataDto { 
    id?: string;
    richContent?: RichContentDto;
    formula?: FormulaDto;
    byteData?: ByteDataDto;
    quantity?: QuantityDto;
    list?: ListDto;
    refIds?: Array<string>;
    refTypes?: Array<string>;
}